Tests for encodePreservingSource
encodePreservingSource tries to additionally provide string_views corresponding
to the sourcetokens where the words came from. To assert this works, the
following method is used:

The alignments provided by encodePreservingSource should match the intended
tokens in the input. The intended tokens are provided as .expected while the
output is generated through encodePreservingSource's alignments.